+++ tests/commands/test_commands.py
@@ -386,10 +386,17 @@ class Environment:
         data_dir.mkdir(exist_ok=True)
         backup_dir = Path(group_dir, "backup")
         backup_dir.mkdir(exist_ok=True)
+        templates_dir = Path(group_dir, "conf/templates")
+        templates_dir.mkdir(exist_ok=True)
+        appcli_dir = Path(group_dir, "conf/templates/appcli")
+        appcli_dir.mkdir(exist_ok=True)
+        context_dir = Path(group_dir, "conf/templates/appcli/context")
+        context_dir.mkdir(exist_ok=True)
 
         return CliContext(
             configuration_dir=conf_dir,
             data_dir=data_dir,
+            application_context_files_dir=None,
             additional_data_dirs=None,
             backup_dir=backup_dir,
             additional_env_variables=None,